updated usecase1

+++ usecase1/cloudera/setupHiveDataExample.sh
@@ -6,12 +6,17 @@ set -e
 
 ROOT_UID=0
 export OUTPUTFILE=pxf_hive_datafile.txt
-
+export HADOOP_USER_NAME=hdfs hive
 # Run as root, of course. (this might not be necessary, because we have to run the script somehow with root anyway)
 if [ "$UID" -ne "$ROOT_UID" ]
 then
   echo "Using $UID to run this script."
-  rm /tmp/$OUTPUTFILE
+
+  if [ -f /data/pxf_examples/$OUTPUTFILE ]
+  then
+    rm /tmp/$OUTPUTFILE
+  fi
+
 
   chown -R gpadmin:gpadmin /tmp/$OUTPUTFILE
   sudo -u hdfs hadoop fs -mkdir -p /data/pxf_examples
@@ -20,7 +25,11 @@ then
 else
   echo "Using root to run this script."
 
-  rm /tmp/$OUTPUTFILE
+  if [ -f /data/pxf_examples/$OUTPUTFILE ]
+  then
+    rm /tmp/$OUTPUTFILE
+  fi
+
   echo -e Prague,Jan,101,4875.33 > /tmp/$OUTPUTFILE
   echo  -e "Prague,Jan,101,4875.33" >> /tmp/$OUTPUTFILE
   echo  -e "Rome,Mar,87,1557.39" >> /tmp/$OUTPUTFILE
@@ -30,8 +39,23 @@ else
   echo  -e "Prague,Dec,333,9894.77" >> /tmp/$OUTPUTFILE
   echo  -e "Bangalore,Jul,271,8320.55" >> /tmp/$OUTPUTFILE
   echo  -e "Beijing,Dec,100,4248.41" >> /tmp/$OUTPUTFILE
-
-  chown -R gpadmin:gpadmin /tmp/$OUTPUTFILE
-  sudo -u hdfs hadoop fs -mkdir -p /data/pxf_examples
-  sudo -u hdfs hadoop fs -put /tmp/$OUTPUTFILE /data/pxf_examples/
 fi
+
+echo "### DROP TABLE IF EXISTS sales_info; "
+beeline -u jdbc:hive2://localhost:10000/default -n scott -p tiger -e "DROP TABLE IF EXISTS sales_info;"
+
+echo "### CREATE TABLE sales_info "
+beeline -u jdbc:hive2://localhost:10000/default -n scott -p tiger -e "CREATE TABLE sales_info (location string, month string,
+        number_of_orders int, total_sales double)
+        ROW FORMAT DELIMITED FIELDS TERMINATED BY ','
+         STORED AS textfile; "
+
+echo "### Upload data into sales_info; "
+beeline -u jdbc:hive2://localhost:10000/default -n scott -p tiger -e "LOAD DATA LOCAL INPATH '/tmp/pxf_hive_datafile.txt'
+        INTO TABLE sales_info;"
+
+echo "### List tables "
+beeline -u jdbc:hive2://localhost:10000/default -n scott -p tiger -e "show tables"
+
+echo "### select * from sales_info "
+beeline -u jdbc:hive2://localhost:10000/default -n scott -p tiger -e "select * from sales_info"
